+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * feat(module)!: use `moduleDependencies` to manipulate options
 * feat(Tooltip): support global content configuration via App tooltip prop
 * feat(DropdownMenu): add `filter` prop
+* feat(InputMenu): add `autocomplete` prop
 * fix(Error): support `status` and `statusText` properties
 * fix(ContentSurround): handle RTL mode
 * fix(Avatar): use resolved size for image width/height